[step1] Answers 일급 콜렉션 리팩토링
